Move Network/Client to Internal/Network/Client
